app/requirements.txt
fastapi==0.95.2
pydantic[email]==1.10.13
uvicorn[standard]==0.23.2
mysql-connector-python==8.3.0
SQLAlchemy>=2.0,<3.0
python-dotenv==1.0.1
PyJWT==2.8.0
python-multipart==0.0.9 
passlib[bcrypt]==1.7.4 